+++ admin/assets/js/admin.js
@@ -12,8 +12,24 @@ var TextInserts = (function ($) {
                             <div class="hb-options-div">\
                                 <span>Box Name</span> <span class="tooltip" tabindex="100" onfocus="TextInserts.displayTooltip(this)" onblur="TextInserts.removeTooltip(this)" data-tooltip="[OPTIONAL] An arbitrary name for this box.">?</span><br>\
                                 <input type="text" name="name" class="name" /><br><br>\
-                                <span>Hook</span> <span class="tooltip" tabindex="100" onfocus="TextInserts.displayTooltip(this)" onblur="TextInserts.removeTooltip(this)" data-tooltip="[REQUIRED] The hook slug where the text/html content will be inserted.<br><br>If you are using the <strong>Thesis 2.1</strong> theme, you can use the hooks detailed here: <a href=&quot;http://diythemes.com/thesis/rtfm/tutorials/hooks/&quot; target=&quot;_blank&quot;>http://diythemes.com/thesis/rtfm/tutorials/hooks/</a><br><br>Examples:<br><ul><li>hook_top_content</li><li>hook_top_header</li></ul>">?</span><br>\
-                                <input type="text" name="hook" class="hook" onkeydown="this.dataset.valid=false" onkeyup="TextInserts.validateHook(this)" data-valid="false" /><br><br>\
+                                <span>Hook</span> <span class="tooltip" tabindex="100" onfocus="TextInserts.displayTooltip(this)" onblur="TextInserts.removeTooltip(this)" data-tooltip="[REQUIRED] The hook slug where the text/html content will be inserted.<br><br>If you are using the <strong>Thesis 2.1</strong> theme, you can use the hooks detailed here: <a href=&quot;http://diythemes.com/thesis/rtfm/tutorials/hooks/&quot; target=&quot;_blank&quot;>Thesis 2 Hook Syntax</a>.<br><br>If you are using the <strong>Genesis</strong> theme, you can use the hooks detailed here: <a href=&quot;http://my.studiopress.com/docs/hook-reference/#structural-action-hooks&quot; target=&quot;_blank&quot;>Structural Action Hooks</a>.<br><br>You can also view all available Genesis/Thesis 2 hooks <a href=&quot;'+ localized.hooks_url +'&quot; target=&quot;_blank&quot;>here</a>.<br><br><i>Note that the hooks beginning with <strong>hook_</strong> in the default list are Thesis hooks.</i>">?</span><br>\
+                                <input type="text" list="hooks-list" name="hook" class="hook" onkeydown="this.dataset.valid=false" onkeyup="TextInserts.validateHook(this)" data-valid="false" /><br><br>\
+                                <datalist id="hooks-list">\
+                                    <option value="hook_before_container">\
+                                    <option value="hook_before_header">\
+                                    <option value="hook_before_columns">\
+                                    <option value="hook_before_content">\
+                                    <option value="hook_before_post_box">\
+                                    <option value="hook_before_sidebar">\
+                                    <option value="hook_before_footer">\
+                                    <option value="genesis_before_header">\
+                                    <option value="genesis_header">\
+                                    <option value="genesis_after_header">\
+                                    <option value="genesis_before_content">\
+                                    <option value="genesis_after_content">\
+                                    <option value="genesis_before_sidebar_widget_area">\
+                                    <option value="genesis_after_sidebar_widget_area">\
+                                </datalist>\
                                 <span>Show in</span> <span class="tooltip" tabindex="100" onfocus="TextInserts.displayTooltip(this)" onblur="TextInserts.removeTooltip(this)" data-tooltip="[REQUIRED] Whether to show this box in posts, pages, both or only in the homepage.">?</span><br>\
                                 <select name="display" class="display">\
                                     <option value="1" selected="selected" >posts &amp; pages</option>\