+++ admin/class-text-inserts-admin.php
@@ -160,7 +160,9 @@ class Text_Inserts_Admin {
 
 		$screen = get_current_screen();
 		if ( $this->plugin_screen_hook_suffix == $screen->id ) {
-			wp_enqueue_script( $this->plugin_slug . '-admin-script', plugins_url( 'assets/js/admin.js', __FILE__ ), array( 'jquery' ), Text_Inserts::VERSION );
+			wp_register_script( $this->plugin_slug . '-admin-script',  plugins_url( 'assets/js/admin.js', __FILE__ ), array( 'jquery' ), Text_Inserts::VERSION );
+			wp_localize_script( $this->plugin_slug . '-admin-script', 'localized', array( 'hooks_url' => plugins_url( 'includes/hooks.html', __FILE__ ) ) );
+			wp_enqueue_script( $this->plugin_slug . '-admin-script' );
 		}
 
 	}